[UPDATE] implementation, tests, eqn

+++ test/test.js
@@ -193,7 +193,7 @@ describe( 'distributions-logistic-mgf', function tests() {
 			expected,
 			i;
 
-		data = new Float32Array( validationData.data );
+		data = new Float64Array( validationData.data );
 
 		expected = new Float64Array( validationData.expected.map( function( d ) {
 			return d === 'Inf' ? Infinity : d;
@@ -217,7 +217,7 @@ describe( 'distributions-logistic-mgf', function tests() {
 			'mu': validationData.mu,
 			's': validationData.s
 		});
-		expected = new Float32Array( validationData.expected.map( function( d ) {
+		expected = new Float64Array( validationData.expected.map( function( d ) {
 			return d === 'Inf' ? Infinity : d;
 		}) );
 		assert.strictEqual( actual, data );
